+++ client/src/components/clientPage/editFormModals/ExitSurveyModal.tsx
@@ -20,59 +20,79 @@ import {
   TableContainer,
   useToast,
 } from "@chakra-ui/react";
-import { useBackendContext } from "../../contexts/hooks/useBackendContext";
-import PrintForm from "../printForm/PrintForm";
-import { SuccessStory } from "../types/SuccessStory";
+import { useBackendContext } from "../../../contexts/hooks/useBackendContext";
+import PrintForm from "../../../printForm/PrintForm";
+import { exitSurvey } from "../types/ExitSurvey";
 
-interface SuccessStoryModalProps {
+interface ExitSurveyModalProps {
   form: { id: number; type: string; title?: string };
   isOpen: boolean;
   onClose: () => void;
 }
 
-export default function SuccessStoryModal({
+export default function ExitSurveyModal({
   form,
   isOpen,
   onClose,
-}: SuccessStoryModalProps) {
+}: ExitSurveyModalProps) {
   const { backend } = useBackendContext();
   const toast = useToast();
 
-  const [storyData, setStoryData] = useState<SuccessStory | null>(null);
+  const [surveyData, setSurveyData] = useState<exitSurvey | null>(null);
   const [loading, setLoading] = useState(false);
   const [isEditing, setIsEditing] = useState(false);
-  const [edits, setEdits] = useState<Partial<SuccessStory>>({});
+  const [edits, setEdits] = useState<Partial<exitSurvey>>({});
 
   const cellHeight = "40px";
 
-  const selectFields: Record<keyof SuccessStory, readonly string[] | undefined> = {
+  const cchRatingOptions: exitSurvey["cch_rating"][] = [
+    "Excellent",
+    "Good",
+    "Fair",
+    "Unsatisfactory",
+  ];
+  const lifeSkillsOptions: exitSurvey["life_skills_rating"][] = [
+    "very helpful",
+    "helpful",
+    "not very helpful",
+    "not helpful at all",
+  ];
+  const cmRatingOptions: exitSurvey["cm_rating"][] = lifeSkillsOptions;
+
+  const selectFields: Record<
+    keyof exitSurvey,
+    readonly string[] | undefined
+  > = {
+    cch_rating: cchRatingOptions,
+    life_skills_rating: lifeSkillsOptions,
+    cm_rating: cmRatingOptions,
     id: undefined,
-    date: undefined,
-    clientId: undefined,
+    cm_id: undefined,
     name: undefined,
-    cmId: undefined,
-    previousSituation: undefined,
-    cchImpact: undefined,
-    whereNow: undefined,
-    tellDonors: undefined,
-    quote: undefined,
-    consent: ["true", "false"],
+    client_id: undefined,
+    site: undefined,
+    date: undefined,
+    program_date_completion: undefined,
+    cch_like_most: undefined,
+    cch_could_be_improved: undefined,
+    life_skills_helpful_topics: undefined,
+    life_skills_offer_topics_in_the_future: undefined,
+    cm_change_about: undefined,
+    cm_most_beneficial: undefined,
+    experience_takeaway: undefined,
+    experience_accomplished: undefined,
+    experience_extra_notes: undefined,
   };
-
-  const toSnake = (s: string): string => {
-    return s.replace(/([A-Z])/g, "_$1").toLowerCase();
-  }
-
 
   useEffect(() => {
     if (!isOpen || !form.id) return;
     setLoading(true);
     backend
-      .get<SuccessStory[]>(`/successStory/${form.id}`)
-      .then((res) => setStoryData(res.data[0]))
+      .get(`/exitSurvey/${form.id}`)
+      .then((res) => setSurveyData(res.data.data[0]))
       .catch((err: any) =>
         toast({
-          title: "Error loading story",
+          title: "Error loading form",
           description: err.message,
           status: "error",
           duration: 3000,
@@ -80,52 +100,52 @@ export default function SuccessStoryModal({
         })
       )
       .finally(() => setLoading(false));
-  }, [isOpen, form.id, backend, toast]);
 
-  // console.log(storyData);
+  }, [isOpen, form.id, backend, toast]);
+  console.log(surveyData);
   const handleEditToggle = () => {
     if (isEditing) setEdits({});
     setIsEditing(!isEditing);
   };
 
-  const handleFieldChange = <K extends keyof SuccessStory>(
+  const handleFieldChange = <K extends keyof exitSurvey>(
     field: K,
-    value: SuccessStory[K] | string
+    value: exitSurvey[K] | string
   ) => {
-    if (field === "id" || field === "cm_id" || field === "client_id") {
-      setEdits((e) => ({ ...e, [field]: Number(value) as SuccessStory[K] }));
-    } else if (field === "consent") {
-      setEdits((e) => ({ ...e, consent: (value === "true") as SuccessStory[K] }));
+    if (["id", "cm_id", "client_id", "site"].includes(field)) {
+      setEdits((e) => ({ ...e, [field]: Number(value) as exitSurvey[K] }));
     } else {
-      setEdits((e) => ({ ...e, [field]: value as SuccessStory[K] }));
+      setEdits((e) => ({ ...e, [field]: value as exitSurvey[K] }));
     }
   };
 
+  // routes and the columns do not follow the same naming convention lol
+  const toCamel = (s: string) =>
+    s.replace(/_([a-z])/g, (_, c) => c.toUpperCase());
+
   const handleSave = async () => {
-    if (!storyData) return;
-    const updated: SuccessStory = { ...storyData, ...edits };
+    if (!surveyData) return;
+    const updated: exitSurvey = { ...surveyData, ...edits };
 
     const payload: Record<string, any> = {};
     for (const key in updated) {
-      payload[toSnake(key)] = (updated as any)[key];
+      payload[toCamel(key)] = (updated as any)[key];
     }
 
-    console.log(payload);
-
     try {
-      await backend.put(`/successStory/${storyData.id}`, payload);
-      setStoryData(updated);
+      await backend.put(`/exitSurvey/${surveyData.id}`, payload);
+      setSurveyData(updated);
       setEdits({});
       setIsEditing(false);
       toast({
-        title: "Story saved successfully",
+        title: "Form saved successfully",
         status: "success",
         duration: 2000,
         isClosable: true,
       });
     } catch (err: any) {
       toast({
-        title: "Error saving story",
+        title: "Error saving form",
         description: err.message,
         status: "error",
         duration: 3000,
@@ -135,7 +155,7 @@ export default function SuccessStoryModal({
   };
 
   if (!isOpen) return null;
-  if (loading || !storyData) {
+  if (loading || !surveyData) {
     return (
       <Modal isOpen={isOpen} onClose={onClose} size="xl">
         <ModalContent>
@@ -148,34 +168,70 @@ export default function SuccessStoryModal({
     );
   }
 
-  const formattedDate = new Date(storyData.date).toLocaleDateString("en-US");
-  const modalTitle = "Success Story – ";
-  const modalDesc = `${form.title || storyData.name} (${formattedDate})`;
+  const formattedDate = new Date(surveyData.date).toLocaleDateString("en-US");
+  const modalTitle = "Form Preview – ";
+  const modalDesc = `${form.title || surveyData.name} (${formattedDate})`;
 
-  const renderField = (field: keyof SuccessStory) => {
-    const original = storyData[field] as any;
-    const edited = edits[field] as any;
+  const renderField = (field: keyof ExitSurvey) => {
+    const original = surveyData[field];
+    const edited = edits[field];
 
     if (!isEditing) {
-      if (field === "date") {
-        return new Date(original).toLocaleDateString("en-US");
-      }
-      if (field === "consent") {
-        return original ? "Yes" : "No";
+      if (field === "date" || field === "program_date_completion") {
+        return new Date(original as string).toLocaleDateString("en-US");
       }
       return String(original);
     }
 
-    if (field === "id" || field === "cm_id" || field === "client_id") {
+    // assuming these fields should be locked
+    if (
+      field === "id" ||
+      field === "cm_id" ||
+      field === "client_id" ||
+      field === "site"
+    ) {
       return (
-        <Box w="100%" h={cellHeight} bg="#EDF2F7" display="flex" alignItems="center" p={3}>
+        <Box
+          w="100%"
+          h={cellHeight}
+          bg="#EDF2F7"
+          display="flex"
+          alignItems="center"
+          p={3}
+        >
           {String(original)}
+
+          {/*
+          // idk if you need to edit these fields or if they should even be allowed to.
+          <Input
+            variant="unstyled"
+            size="sm"
+            w="100%"
+            h={cellHeight}
+            px={3}
+            m={0}
+            border="1px solid"
+            borderColor="#3182CE"
+            type="number"
+            value={String(edited ?? original)}
+            onChange={(e) =>
+              handleFieldChange(field, e.target.value)
+            }
+          />
+          */}
         </Box>
       );
     }
 
     return (
-      <Box w="100%" h={cellHeight} bg="#EDF2F7" display="flex" alignItems="center" p={0}>
+      <Box
+        w="100%"
+        h={cellHeight}
+        bg="#EDF2F7"
+        display="flex"
+        alignItems="center"
+        p={0}
+      >
         {selectFields[field] ? (
           <Box
             w="100%"
@@ -183,6 +239,7 @@ export default function SuccessStoryModal({
             bg="#EDF2F7"
             border="1px solid"
             borderColor="#3182CE"
+            borderRadius="0"
             px={3}
             display="flex"
             alignItems="center"
@@ -196,12 +253,12 @@ export default function SuccessStoryModal({
               lineHeight={cellHeight}
               _focus={{ boxShadow: "none" }}
               sx={{ appearance: "none" }}
-              value={edited !== undefined ? String(edited) : String(original)}
+              value={(edited ?? original) as string}
               onChange={(e) => handleFieldChange(field, e.target.value)}
             >
               {selectFields[field]!.map((opt) => (
                 <option key={opt} value={opt}>
-                  {opt === "true" ? "Yes" : "No"}
+                  {opt}
                 </option>
               ))}
             </Select>
@@ -216,34 +273,52 @@ export default function SuccessStoryModal({
             m={0}
             border="1px solid"
             borderColor="#3182CE"
-            type={field === "date" ? "date" : "text"}
+            type={
+              field === "date" || field === "program_date_completion"
+                ? "date"
+                : "text"
+            }
             value={
               edited !== undefined
                 ? String(edited)
-                : field === "date"
+                : field === "date" || field === "program_date_completion"
                 ? (original as string)
                 : String(original)
             }
             onChange={(e) => handleFieldChange(field, e.target.value)}
-            onKeyDown={field === "date" ? (e) => e.preventDefault() : undefined}
+            onKeyDown={
+              field === "date" ? (e) => e.preventDefault() : undefined
+            }
           />
         )}
       </Box>
     );
   };
 
-  const rows: Array<{ field: keyof SuccessStory; label: string }> = [
+
+  const rows: Array<{ field: keyof exitSurvey; label: string }> = [
     { field: "id", label: "ID" },
+    { field: "cm_id", label: "Case Manager ID" },
     { field: "date", label: "Date" },
     { field: "name", label: "Name" },
-    { field: "clientId", label: "Client ID" },
-    { field: "cmId", label: "Case Manager ID" },
-    { field: "previousSituation", label: "Previous Situation" },
-    { field: "cchImpact", label: "CCH Impact" },
-    { field: "whereNow", label: "Where Now" },
-    { field: "tellDonors", label: "Message to Donors" },
-    { field: "quote", label: "Quote" },
-    { field: "consent", label: "Consent" },
+    { field: "client_id", label: "Client ID" },
+    { field: "site", label: "Site" },
+    { field: "program_date_completion", label: "Program Date Completion" },
+    { field: "cch_rating", label: "CCH Rating" },
+    { field: "cch_like_most", label: "What did you like most?" },
+    { field: "cch_could_be_improved", label: "CCH Could Be Improved" },
+    { field: "life_skills_rating", label: "Life Skills Rating" },
+    { field: "life_skills_helpful_topics", label: "Life Skills Helpful Topics" },
+    {
+      field: "life_skills_offer_topics_in_the_future",
+      label: "Life Skills Offer Topics in the Future",
+    },
+    { field: "cm_rating", label: "CM Rating" },
+    { field: "cm_change_about", label: "CM Change About" },
+    { field: "cm_most_beneficial", label: "CM Most Beneficial" },
+    { field: "experience_takeaway", label: "Experience Takeaway" },
+    { field: "experience_accomplished", label: "Experience Accomplished" },
+    { field: "experience_extra_notes", label: "Experience Extra Notes" },
   ];
 
   return (
@@ -275,7 +350,7 @@ export default function SuccessStoryModal({
             {!isEditing ? (
               <>
                 <Button variant="outline" colorScheme="blue" onClick={handleEditToggle}>
-                  Edit Story
+                  Edit Form
                 </Button>
                 <Box
                   display="inline-block"
@@ -287,7 +362,7 @@ export default function SuccessStoryModal({
                   cursor="pointer"
                   _hover={{ backgroundColor: "blue.600" }}
                 >
-                  <PrintForm formType="Success Story" formId={form.id} />
+                  <PrintForm formType="Exit Survey" formId={form.id} />
                 </Box>
               </>
             ) : (
